+++ packages/cryptoplease_link/lib/src/constants.dart
@@ -21,12 +21,19 @@ const minimumSwapFee = 10000; // 0.01 USDC
 const solDecimals = 1000000000;
 const usdcDecimals = 1000000;
 
+const int tokenProgramRent = 2039280;
+const lamportsPerSignature = 5000;
+
 final devnetUsdc = Ed25519HDPublicKey.fromBase58(
   '4zMMC9srt5Ri5X14GAgXhaHii3GnPAEERYPJgZJDncDU',
 );
 
 final mainnetUsdc = Ed25519HDPublicKey.fromBase58(
   'EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v',
+);
+
+final wrappedSol = Ed25519HDPublicKey.fromBase58(
+  'So11111111111111111111111111111111111111112',
 );
 
 final devnetPlatformMnemonic =